updated controllers package structure; added sum to cart
