--- src/main/java/vlad/petrovskyi/internetshop/controllers/DeleteUserController.java
@@ -1,24 +1,0 @@
-package vlad.petrovskyi.internetshop.controllers;
-
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import vlad.petrovskyi.internetshop.lib.Injector;
-import vlad.petrovskyi.internetshop.service.UserService;
-
-public class DeleteUserController extends HttpServlet {
-
-    private static final Injector injector = Injector.getInstance("vlad.petrovskyi.internetshop");
-    private final UserService userService = (UserService) injector.getInstance(UserService.class);
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        String userId = req.getParameter("user_id");
-        userService.delete(Long.valueOf(userId));
-
-        resp.sendRedirect(req.getContextPath() + "/getAllUsers");
-    }
-}
