+++ src/main/java/vlad/petrovskyi/internetshop/controllers/db/products/AllProductsFromDbController.java
@@ -1,29 +1,27 @@
-package vlad.petrovskyi.internetshop.controllers;
+package vlad.petrovskyi.internetshop.controllers.db.products;
 
 import java.io.IOException;
+import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import vlad.petrovskyi.internetshop.lib.Injector;
+import vlad.petrovskyi.internetshop.model.Product;
 import vlad.petrovskyi.internetshop.service.ProductService;
-import vlad.petrovskyi.internetshop.service.ShoppingCartService;
 
-public class AddToCartController extends HttpServlet {
-    private static final Long USER_ID = 1L;
+public class AllProductsFromDbController extends HttpServlet {
 
     private static final Injector injector = Injector.getInstance("vlad.petrovskyi.internetshop");
     private final ProductService productService =
             (ProductService) injector.getInstance(ProductService.class);
-    private final ShoppingCartService shoppingCartService =
-            (ShoppingCartService) injector.getInstance(ShoppingCartService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        shoppingCartService.addProduct(
-                shoppingCartService.getByUserId(USER_ID),
-                productService.get(Long.valueOf(req.getParameter("product_id"))));
-        resp.sendRedirect(req.getContextPath() + "/allProductsToBuy");
+        List<Product> productList = productService.getAll();
+
+        req.setAttribute("products", productList);
+        req.getRequestDispatcher("/WEB-INF/views/products/allProductsEdit.jsp").forward(req, resp);
     }
 }
