+++ src/main/java/vlad/petrovskyi/internetshop/controllers/db/products/DeleteProductFromDbController.java
@@ -1,16 +1,14 @@
-package vlad.petrovskyi.internetshop.controllers;
+package vlad.petrovskyi.internetshop.controllers.db.products;
 
 import java.io.IOException;
-import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import vlad.petrovskyi.internetshop.lib.Injector;
-import vlad.petrovskyi.internetshop.model.Product;
 import vlad.petrovskyi.internetshop.service.ProductService;
 
-public class AllProductsFromDbController extends HttpServlet {
+public class DeleteProductFromDbController extends HttpServlet {
 
     private static final Injector injector = Injector.getInstance("vlad.petrovskyi.internetshop");
     private final ProductService productService =
@@ -19,9 +17,7 @@ public class AllProductsFromDbController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<Product> productList = productService.getAll();
-
-        req.setAttribute("products", productList);
-        req.getRequestDispatcher("/WEB-INF/views/products/allProductsEdit.jsp").forward(req, resp);
+        productService.delete(Long.valueOf(req.getParameter("product_id")));
+        resp.sendRedirect(req.getContextPath() + "/allProductsFromDb");
     }
 }
