+++ src/main/java/vlad/petrovskyi/internetshop/controllers/db/users/DeleteUserController.java
@@ -1,4 +1,4 @@
-package vlad.petrovskyi.internetshop.controllers;
+package vlad.petrovskyi.internetshop.controllers.db.users;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -6,18 +6,19 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import vlad.petrovskyi.internetshop.lib.Injector;
-import vlad.petrovskyi.internetshop.service.ProductService;
+import vlad.petrovskyi.internetshop.service.UserService;
 
-public class DeleteProductFromDbController extends HttpServlet {
+public class DeleteUserController extends HttpServlet {
 
     private static final Injector injector = Injector.getInstance("vlad.petrovskyi.internetshop");
-    private final ProductService productService =
-            (ProductService) injector.getInstance(ProductService.class);
+    private final UserService userService = (UserService) injector.getInstance(UserService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        productService.delete(Long.valueOf(req.getParameter("product_id")));
-        resp.sendRedirect(req.getContextPath() + "/allProductsFromDb");
+        String userId = req.getParameter("user_id");
+        userService.delete(Long.valueOf(userId));
+
+        resp.sendRedirect(req.getContextPath() + "/getAllUsers");
     }
 }
