+++ src/main/java/vlad/petrovskyi/internetshop/controllers/user/UserPageController.java
@@ -1,4 +1,4 @@
-package vlad.petrovskyi.internetshop.controllers;
+package vlad.petrovskyi.internetshop.controllers.user;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
